+++ react/utils/shared.ts
@@ -25,11 +25,13 @@ export const statusTagControl = (
 ): {
   label: string
   palette: PaletteColors
+  size: 'normal' | 'large'
 } => {
   if (status === 'order-created') {
     return {
       label: intl.formatMessage(messages.orderStatusCreatedLabel),
       palette: 'gray',
+      size: 'large',
     }
   }
 
@@ -37,6 +39,7 @@ export const statusTagControl = (
     return {
       label: intl.formatMessage(messages.orderStatusPendingLabel),
       palette: 'gray',
+      size: 'large',
     }
   }
 
@@ -44,6 +47,7 @@ export const statusTagControl = (
     return {
       label: intl.formatMessage(messages.orderStatusPaidLabel),
       palette: undefined,
+      size: 'large',
     }
   }
 
@@ -51,11 +55,13 @@ export const statusTagControl = (
     return {
       label: intl.formatMessage(messages.orderStatusInvoicedLabel),
       palette: 'green',
+      size: 'large',
     }
   }
 
   return {
     label: intl.formatMessage(messages.orderStatusCancelLabel),
     palette: 'red',
+    size: 'large',
   }
 }